refactor: JWT 별도 디렉토리 생성 및 auth 도메인 생성
