+++ src/main/java/org/example/namelesschamber/domain/user/controller/UserController.java
@@ -5,7 +5,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.example.namelesschamber.common.response.ApiResponse;
-import org.example.namelesschamber.common.security.SecurityUtils;
+import org.example.namelesschamber.domain.auth.core.SecurityUtils;
 import org.example.namelesschamber.domain.user.dto.request.NicknameRequestDto;
 import org.example.namelesschamber.domain.user.dto.response.UserInfoResponseDto;
 import org.example.namelesschamber.domain.user.service.UserService;
